spin getOptions
